fix: Update login response format to match API specifications

+++ src/main/java/com/shotty/shotty/dto/common/ResponseDto.java
@@ -1,7 +1,7 @@
 package com.shotty.shotty.dto.common;
 
 public record ResponseDto<T>(
-        short code,
+        int code,
         String statusMsg,
         T data
 ) {
